requirements.txt
flask==3.1.0
Flask-WTF==1.2.2
wtforms==3.2.1
pandas==2.2.3
pyyaml==6.0.2
python-box==7.3.0
ensure==1.0.4
werkzeug==3.1.0
ipykernel==6.29.5
scikit-learn==1.6.0
matplotlib==3.9.4
seaborn==0.13.2
nltk==3.9.1
wordcloud==1.9.4
numpy==2.0.2
datasets==3.2.0
symspellpy==6.7.8
transformers==4.48.0
torch==2.5.1
gunicorn==23.0.0
dnspython==2.7.0
-e .